+++ Insurance-final-project/Models/PolicyAccount.cs
@@ -16,8 +16,8 @@ namespace Insurance_final_project.Models
         public Guid CustomerId { get; set; }
         public Customer Customer { get; set; }
 
-        public decimal CoverageAmount { get; set; }
-        public decimal TotalAmountPaid { get; set; }
+        public double CoverageAmount { get; set; }
+        public double TotalAmountPaid { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public string InstallmentType { get; set; } // Monthly, Quarterly, etc.
@@ -27,6 +27,6 @@ namespace Insurance_final_project.Models
         public Agent Agent { get; set; }
 
         public string Status { get; set; } // Open/Closed
-        public decimal AgentCommission { get; set; } // Calculated based on installment payments and commission percentage
+        public double AgentCommission { get; set; } // Calculated based on installment payments and commission percentage
     }
 }
